ai_learn/week6_eval/requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
pyyaml>=6.0
tqdm>=4.66
rapidfuzz>=3.5
